Create NUTS UK shapefiles
This folder contains the Uk shapefiles with the EU NUTS classification (Nomenclature of territorial units for statistics). NUTS is a hierarchical system for dividing up the economic territory of the EU, including the UK for the purpose of the collection, development and harmonisation of European regional statistics. It has 3 levels, among which:
         NUTS 1: major socio-economic regions
         NUTS 2: basic regions for the application of regional policies
         NUTS 3: small regions for specific diagnoses